ASTDiff: Introduce SingleMonacoContent
